+++ client/src/app/components/ImpactsTextInput.tsx
@@ -3,12 +3,12 @@ import { useDispatch } from 'react-redux';
 // ---
 import { Tooltip } from 'app/components/Tooltip';
 import { useTypedSelector } from 'app/redux/index';
-import { calculateHourlyEnergyProfile } from 'app/redux/reducers/eere';
+import { calculateHourlyEnergyProfile } from 'app/redux/reducers/impacts';
 import type {
   EnergyEfficiencyFieldName,
   RenewableEnergyFieldName,
   ElectricVehiclesFieldName,
-} from 'app/redux/reducers/eere';
+} from 'app/redux/reducers/impacts';
 
 export function ImpactsTextInput(props: {
   className?: string;
