+++ client/src/app/redux/index.ts
@@ -6,7 +6,7 @@ import api from 'app/redux/reducers/api';
 import panel from 'app/redux/reducers/panel';
 import geography from 'app/redux/reducers/geography';
 import transportation from 'app/redux/reducers/transportation';
-import impacts from 'app/redux/reducers/eere';
+import impacts from 'app/redux/reducers/impacts';
 import results from 'app/redux/reducers/results';
 import downloads from 'app/redux/reducers/downloads';
 import monthlyEmissions from 'app/redux/reducers/monthlyEmissions';
